\name{Rhy-package}
\alias{Rhy-package}
\alias{Rhy}
\docType{package}
\title{
Haoyu's Implementation of Useful Toolkits
}
\description{
This package will contain my implementation of many useful toolkits related
to linear and convex optimization, machine learning, data mining, etc.
}
\details{
\tabular{ll}{
Package: \tab Rhy\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2013-10-25\cr
License: \tab GPL (>= 2)\cr
}
This package is currently under development.
}
\author{
Haoyu Tan <haoyutan@gmail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
